apis: replace unsigned integers in CRD fields with signed integers, add validation
This commit replaces the unsigned integers in the HTTPProxy and IngressRoute CRDs
with signed integers. Unsigned integers are discouraged due to API compatibility
issues. Validation rules have been added to the fields to ensure that the value
provided is always positive.

Fixes #2123

+++ internal/dag/policy.go
@@ -32,7 +32,7 @@ func retryPolicy(rp *projcontour.RetryPolicy) *RetryPolicy {
 	perTryTimeout, _ := time.ParseDuration(rp.PerTryTimeout)
 	return &RetryPolicy{
 		RetryOn:       "5xx",
-		NumRetries:    max(1, rp.NumRetries),
+		NumRetries:    max(1, uint32(rp.NumRetries)),
 		PerTryTimeout: perTryTimeout,
 	}
 }
@@ -156,8 +156,8 @@ func ingressrouteHealthCheckPolicy(hc *ingressroutev1.HealthCheck) *HTTPHealthCh
 		Host:               hc.Host,
 		Interval:           time.Duration(hc.IntervalSeconds) * time.Second,
 		Timeout:            time.Duration(hc.TimeoutSeconds) * time.Second,
-		UnhealthyThreshold: hc.UnhealthyThresholdCount,
-		HealthyThreshold:   hc.HealthyThresholdCount,
+		UnhealthyThreshold: uint32(hc.UnhealthyThresholdCount),
+		HealthyThreshold:   uint32(hc.HealthyThresholdCount),
 	}
 }
 
@@ -170,8 +170,8 @@ func httpHealthCheckPolicy(hc *projcontour.HTTPHealthCheckPolicy) *HTTPHealthChe
 		Host:               hc.Host,
 		Interval:           time.Duration(hc.IntervalSeconds) * time.Second,
 		Timeout:            time.Duration(hc.TimeoutSeconds) * time.Second,
-		UnhealthyThreshold: hc.UnhealthyThresholdCount,
-		HealthyThreshold:   hc.HealthyThresholdCount,
+		UnhealthyThreshold: uint32(hc.UnhealthyThresholdCount),
+		HealthyThreshold:   uint32(hc.HealthyThresholdCount),
 	}
 }
